refactor the dashboard.py
